pyproject.toml
[tool.poetry]
name = "valorant-analysis"
version = "0.1.0"
description = "Analyzing Valorant match data."
authors = ["Nick Anthony <nanthony007@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^1.3.3"
PyYAML = "^6.0"
scipy = "^1.9.0"
scikit-learn = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
